Updated more PNGs
NUG_multiple_plots_along_y.png and NUG_triangular_grid_ICON_640.png are
both slightly different on Linux and Mac, but are similar enough for
"compare -metric DSSIM" to consider them to be "similar". The copies of
these PNGs included previously were generated on Mac, the replacements
are generated on Linux.